+++ jobrunner/signals.py
@@ -1,5 +1,8 @@
 from django.db.models.signals import post_save
 from django.dispatch import receiver
+from django.conf import settings
+
+import json
 
 from . import tasks, models
 
@@ -8,4 +11,9 @@ from . import tasks, models
 def start_job(sender, instance, created, **kwargs):
     if created:
         id_ = str(instance.id)
-        tasks.execute.delay(id_)
+        inputs = json.loads(instance.inputs)
+        immediate = inputs.get('immediate', False)
+        if immediate and settings.ALLOW_BLOCKING_BMDS_REQUESTS:
+            instance.try_execute()
+        else:
+            tasks.try_execute.delay(id_)