+++ bmds_server/settings/production.py
@@ -35,4 +35,5 @@ LOGGING['handlers']['file'].update(
 
 BROKER_URL = os.environ['CELERY_BROKER_URL']
 CELERY_RESULT_BACKEND = os.environ['CELERY_RESULT_BACKEND']
-ALLOW_BLOCKING_BMDS = bool(os.environ.get('ALLOW_BLOCKING_BMDS', 'False') == 'True')
+
+ALLOW_BLOCKING_BMDS_REQUESTS = bool(os.environ.get('ALLOW_BLOCKING_BMDS_REQUESTS', 'False') == 'True')