+++ HISTORY.rst
@@ -6,6 +6,7 @@ v0.2.0 (NEXT)
 -------------------
 
 * For continuous summary datasets, rename ``responses`` array to ``means``
+* Added quickstart for using R
 
 
 v0.1.0 (2016-12-05)