+++ HISTORY.rst
@@ -8,6 +8,7 @@ v0.2.0 (NEXT)
 * For continuous summary datasets, rename ``responses`` array to ``means``
 * Added quickstart for using R
 * Added the ability to enter continuous individual datasets
+* Added new option on requests to allow ``immediate`` - if True and enabled on the server, a user can request an immediate response to a job, and therefore polling for results is no-longer required. False by default for server performance.
 
 
 v0.1.0 (2016-12-05)